pyproject.toml
[project]
name = "datastats-gcp-jobs-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests",
    "loguru",
    "beautifulsoup4",
    "python-dotenv",
    "google-cloud-storage>=1.17.0",
    "pg8000",
    "google-cloud-logging"
]
